(200427)Ahna_FPS #1

+++ Client/CorruptLab#2/CorruptLab/Shader_BillBoard.cpp
@@ -445,8 +445,8 @@ void CSoftParticleShader::BuildObjects(ID3D12Device* pd3dDevice, ID3D12GraphicsC
 	{
 		for (int j = 50; j < 400; j+=80)
 		{
-			float fHeight = pTerrain->GetHeight(i, j) + 30;
-			pNoise = new CObjectFog(pd3dDevice, pd3dCommandList, pd3dGraphicsRootSignature, XMFLOAT3(i, fHeight, j), this);  //object
+			float fHeight = pTerrain->GetHeight((float)i, (float)j) + 30;
+			pNoise = new CObjectFog(pd3dDevice, pd3dCommandList, pd3dGraphicsRootSignature, XMFLOAT3((float)i, (float)fHeight, (float)j), this);  //object
 			pNoise->GenerateShaderDistortionBuffer();
 			m_pFogObjects[w] = pNoise; w++;
 		}
